状態管理やメソッドを子コンポーネントに移行
